backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
email-validator==2.1.1
python-dotenv==1.0.1
